refactor(scroll): collapse --until onto the one route that runs it
