+++ scripts/layering/package-boundaries.test.ts
@@ -418,7 +418,6 @@ test('the real tree parses, declares, and passes R11', () => {
     '@agent-device/capture-kit/screenshot-diff-pixels',
     '@agent-device/capture-kit/screenshot-overlay',
     '@agent-device/capture-kit/scroll-edge-state',
-    '@agent-device/capture-kit/scroll-until-visible',
     '@agent-device/capture-kit/snapshot-chrome',
     '@agent-device/capture-kit/snapshot-desktop-projection',
     '@agent-device/capture-kit/snapshot-desktop-surface',
@@ -653,7 +652,6 @@ test('the real tree parses, declares, and passes R11', () => {
     '@agent-device/selectors/interaction-touch-point',
     '@agent-device/selectors/parameterized-recorded-fill',
     '@agent-device/selectors/press-retarget',
-    '@agent-device/selectors/scroll-until-match',
     '@agent-device/selectors/selector-pipeline',
     '@agent-device/selectors/selector-pipeline-policy',
     '@agent-device/selectors/snapshot-geometry-fixtures',
